Enable zeitwerk autoloader
this is default for rails 6.0 config. nesting of middlewares was
updated to match expected class names by zeitwerk.

requried to pass `rake zeitwerk:check

```
NameError: uninitialized constant Hostess
```

lib is part of autoload_path. we need to explicity exclude folder
from lib which don't need to be autoloated (initializers/zeitwerk.rb).
