+++ config/environments/development.rb
@@ -1,4 +1,4 @@
-require_relative "../../lib/middleware/hostess"
+require_relative "../../lib/gemcutter/middleware/hostess"
 require "active_support/core_ext/integer/time"
 
 Rails.application.configure do
@@ -66,7 +66,7 @@ Rails.application.configure do
   # Raises error for missing translations.
   # config.i18n.raise_on_missing_translations = true
 
-  config.middleware.use Hostess
+  config.middleware.use Gemcutter::Middleware::Hostess
   # Annotate rendered view with file names.
   # config.action_view.annotate_rendered_view_with_filenames = true
